fix: address build test fail (wild guess)
