+++ tests/cucumber/support/step_definitions/Then I see DefaultImportModalDialog.js
@@ -1,0 +1,7 @@
+import { materialDesignerPage } from "../widgets/material_designer_page";
+
+export default function () {
+    this.Then(/^I see DefaultImportModalDialog$/, () => {
+        materialDesignerPage.designerWidget.defaultImportModalDialog.isVisible();
+    });
+}
