+++ tests/cucumber/support/step_definitions/Then I see the Add button.js
@@ -1,0 +1,7 @@
+import { materialDesignerPage } from "../widgets/material_designer_page";
+
+export default function () {
+    this.Then(/^I see the Add button$/, () => {
+        materialDesignerPage.designerWidget.defaultImportModalDialog.addButtonExists();
+    });
+}
