+++ tests/cucumber/support/step_definitions/Then the DefaultImportModalDialog should be closed.js
@@ -1,0 +1,7 @@
+import { materialDesignerPage } from "../widgets/material_designer_page";
+
+export default function () {
+    this.Then(/^the DefaultImportModalDialog should be closed$/, () => {
+        materialDesignerPage.designerWidget.defaultImportModalDialog.waitForDisappear();
+    });
+}
