+++ tests/cucumber/support/step_definitions/When I cancel import.js
@@ -1,0 +1,7 @@
+import { materialDesignerPage } from "../widgets/material_designer_page";
+
+export default function () {
+    this.When(/^I cancel import$/, () => {
+        materialDesignerPage.designerWidget.defaultImportModalDialog.cancel();
+    });
+}
