+++ tests/cucumber/support/step_definitions/When I click the Submit button.js
@@ -1,0 +1,7 @@
+import { materialDesignerPage } from "../widgets/material_designer_page";
+
+export default function () {
+    this.When(/^I click the Submit button$/, () => {
+        materialDesignerPage.designerWidget.defaultImportModalDialog.submit();
+    });
+}
