+++ tests/cucumber/support/step_definitions/When I upload files.js
@@ -1,0 +1,9 @@
+import { parseTable } from "../utils/table";
+import { materialDesignerPage } from "../widgets/material_designer_page";
+
+export default function () {
+    this.When(/^I upload files$/, function (table) {
+        const filenames = parseTable(table, this);
+        materialDesignerPage.designerWidget.defaultImportModalDialog.uploadFiles(filenames);
+    });
+}
